move messageName constants to enum
