+++ src/main/java/uk/gov/hmcts/reform/wacaseeventhandler/services/HandlerConstants.java
@@ -1,0 +1,28 @@
+package uk.gov.hmcts.reform.wacaseeventhandler.services;
+
+public enum HandlerConstants {
+
+    TASK_CANCELLATION("wa-task-cancellation", "cancelTasks"),
+    TASK_INITIATION("wa-task-initiation", "createTaskMessage");
+
+    private final String tableName;
+    private final String messageName;
+
+    HandlerConstants(String tableName, String messageName) {
+        this.tableName = tableName;
+        this.messageName = messageName;
+    }
+
+    public String getTableName() {
+        return tableName;
+    }
+
+    public String getMessageName() {
+        return messageName;
+    }
+
+    public String getTableKey(String jurisdictionId, String caseTypeId) {
+        return tableName + "-" + jurisdictionId + "-" + caseTypeId;
+    }
+
+}
